Expose the new API as `retryWhen(Retry)` despite ambiguity
We'll deprecate all the retryBackoff methods as well as the original
retryWhen (based on Throwable companion publisher) in order to introduce
the new `RetrySignal` based signature. The use of `Retry` explicit type
lifts any ambiguity when using the Builder, but using a lambda instead
will raise some ambiguity at call site. We deem that acceptable given
that the migration is quite easy (turn `e -> whatever(e)` to `(Retry)
rs -> whatever(rs.failure())`), and that the method that generates
ambiguity is deprecated and slated for removal.