Meld the xxRetryFunction with the Builder, have 2 builders
The `build()` step seems useless, so we can have 2 builders dedicated to simple
and backoff case respectively, and have these builders directly implement the
Retry interface.

With concrete builders, the template copy on write nature is maintained.
There is some amount of API duplication, but that is a small price to
pay since copy constructor and return type consistency force us to have
all the methods implemented in each concrete class anyway (including the
common base).