Remove the `Supplier` from retryWhen
The resulting ambiguity is the least likely to cause trouble for users,
as the operator is already considered advanced and the ambiguity only
raises when providing the Function as a lambda. reactor-extra Retry
builder for instance shouldn't raise an ambiguity.

Also polish the javadoc a bit and migrate MonoRetryWhenTest to
undeprecated APIs.